code/argali_tether/requirements.txt
pyserial
coverage
